Automatical creation of `.dir_info`, commands and Saving/restoring design change (#38)
* automatically make sekai valid to a certain extent

* tap command, make files and directories

* restore beginning, save every time

This means restore will be created just once in the beginning and every
time save will be get restored

* copy, delete, exit and more commands

* help text add

* add prompter for force overwrite

* Update src/utils/valid_sekai.rs


* simplify and rename error of sekai not valid

* Update src/commands/copy.rs


* remove unnecessary line of check_dir_info_exists


---------

+++ src/utils/restore_comp.rs
@@ -17,7 +17,7 @@ fn generate_temp_path(usage: &str, root_path: &PathBuf) -> PathBuf {
     PathBuf::from(format!("/tmp/deemak-{}-{:x}", usage, hash))
 }
 
-pub fn backup_sekai(usage: &str, root_path: &PathBuf) -> std::io::Result<()> {
+pub fn backup_sekai(usage: &str, root_path: &PathBuf) -> std::io::Result<String> {
     let dir_info_path = root_path.join(".dir_info");
     fs::create_dir_all(&dir_info_path)?;
 
@@ -31,6 +31,11 @@ pub fn backup_sekai(usage: &str, root_path: &PathBuf) -> std::io::Result<()> {
             ));
         }
     };
+
+    // if `restore_me` already exists, then do not recreate it
+    if usage == "restore" && dir_info_path.join(RESTORE_FILE).exists() {
+        return Ok("Restore file already exists, skipping creation.".to_string());
+    }
 
     let file = File::create(&backup_file)?;
     let mut encoder = ZlibEncoder::new(file, Compression::best());
@@ -59,7 +64,7 @@ pub fn backup_sekai(usage: &str, root_path: &PathBuf) -> std::io::Result<()> {
     }
 
     encoder.finish()?;
-    Ok(())
+    Ok(format!("Backup created at {:?}", backup_file))
 }
 
 pub fn restore_sekai(usage: &str, root_path: &PathBuf) -> std::io::Result<()> {
@@ -107,6 +112,7 @@ pub fn restore_sekai(usage: &str, root_path: &PathBuf) -> std::io::Result<()> {
     let mut archive = tar::Archive::new(decoder);
     archive.unpack(root_path)?;
 
+    // Clean up temporary file
     fs::remove_file(temp_path)?;
     Ok(())
 }
